+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add env property to steps to control which steps have access to
   `SRVC_TOKEN`.
+- Use `SRVC_TOKEN` to set Authorization header in generator embedded step.
 
 ## v0.16.0 (2023-03-31)